requirements.txt
anyio==3.6.2
asttokens==2.2.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bertopic==0.14.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
Cython==0.29.34
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
Faker==18.6.2
fastapi==0.95.1
filelock==3.12.0
fsspec==2023.4.0
greenlet==2.0.2
h11==0.14.0
hdbscan==0.8.29
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
llvmlite==0.40.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numba==0.57.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pip==23.1.2
platformdirs==3.5.0
plotly==5.14.1
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.15.1
pynndescent==0.5.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==305.1
PyYAML==6.0
pyzmq==25.0.2
regex==2023.5.5
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==67.7.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
stack-data==0.6.2
starlette==0.26.1
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.0
torchvision==0.15.1
tornado==6.3
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
typing_extensions==4.5.0
tzdata==2023.3
umap-learn==0.5.3
urllib3==2.0.2
uvicorn==0.22.0
wcwidth==0.2.6
wheel==0.40.0
zipp==3.15.0
